修改hilog落盘任务抓取逻辑

+++ plugins/ohos/src/ohos/environment/device.py
@@ -433,13 +433,21 @@ class Device(IDevice):
     def stop_hilog_task(self, log_name):
         cmd = "hilog -w stop"
         out = self.execute_shell_command(cmd)
-        # 把hilog文件夹下所有文件拉出来 由于hdc不支持整个文件夹拉出只能采用先压缩再拉取文件
-        cmd = "cd /data/log/hilog && tar -zcvf /data/log/hilog_{}.tar.gz *".format(log_name)
-        out = self.execute_shell_command(cmd)
-        LOG.info("Execute command: {}, result is {}".format(cmd, out))
-        self.pull_file("/data/log/hilog_{}.tar.gz".format(log_name), "{}/log/".format(self._device_log_path))
-        cmd = "rm -rf /data/log/hilog_{}.tar.gz".format(log_name)
-        out = self.execute_shell_command(cmd)
+        self.pull_file("/data/log/hilog/", "{}/log/".format(self._device_log_path))
+        try:
+            os.rename("{}/log/hilog".format(self._device_log_path),
+                      "{}/log/{}_hilog".format(self._device_log_path, log_name))
+        except Exception as e:
+            self.log.error("Rename hilog folder {}_hilog failed. error: {}".format(log_name, e))
+            # 把hilog文件夹下所有文件拉出来 由于hdc不支持整个文件夹拉出只能采用先压缩再拉取文件
+            cmd = "cd /data/log/hilog && tar -zcvf /data/log/hilog_{}.tar.gz *".format(log_name)
+            out = self.execute_shell_command(cmd)
+            LOG.info("Execute command: {}, result is {}".format(cmd, out))
+            if "No space left on device" not in out:
+                self.pull_file("/data/log/{}_hilog.tar.gz".format(log_name),
+                               "{}/log/".format(self._device_log_path))
+                cmd = "rm -rf /data/log/hilog_{}.tar.gz".format(log_name)
+                out = self.execute_shell_command(cmd)
         # 获取crash日志
         self.start_get_crash_log(log_name)
